requirements.txt
haystack-ai>=2.24.1
python-dotenv
click
aiohttp
trafilatura
jinja2
